// app/generated/api/models/accountAttributeType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pricing Backend API
 * REST API for managing branches, regions, products, and pricing plans.
 * OpenAPI spec version: 1.0.0
 */

export type AccountAttributeType = typeof AccountAttributeType[keyof typeof AccountAttributeType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AccountAttributeType = {
  TEXT: 'TEXT',
  DECIMAL: 'DECIMAL',
  INTEGER: 'INTEGER',
  DATE: 'DATE',
  BOOLEAN: 'BOOLEAN',
} as const;