// app/generated/api/models/accountAttributeOption.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pricing Backend API
 * REST API for managing branches, regions, products, and pricing plans.
 * OpenAPI spec version: 1.0.0
 */
import type { Id } from './id';
import type { AccountAttributeCode } from './accountAttributeCode';
import type { AccountAttributeName } from './accountAttributeName';
import type { AccountAttributeType } from './accountAttributeType';

export interface AccountAttributeOption {
  id: Id;
  code: AccountAttributeCode;
  name: AccountAttributeName;
  type: AccountAttributeType;
}