// app/generated/api/models/reasonCondition.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pricing Backend API
 * REST API for managing branches, regions, products, and pricing plans.
 * OpenAPI spec version: 1.0.0
 */
import type { Id } from './id';
import type { ReasonOperator } from './reasonOperator';
import type { ReasonConditionValue } from './reasonConditionValue';

export interface ReasonCondition {
  attributeId: Id;
  operator: ReasonOperator;
  value: ReasonConditionValue;
}